Create README.md
RSI is a most useful technical indicator for technical analysis of the stock. This code tests the RSI based logic on 5 Indian stocks and measures the performance in terms of profit/loss and # of trades done. Idea is to download the last 5 year data and generate signals like when RSI<40 then buy and when RSI>60 sell. Resuklts arent too great but indeed promising.